Fix [object Object] in Widget Permissions
Due to two-step translations we could sometimes end up in a case where
instead of a string we'd get a React Node wrapper, which tried being
translated for the second time. This commit solves this problem by
verifying if we have passed both Tags and Variables.

Fixes https://github.com/vector-im/element-web/issues/18384